requirements.txt
asgiref==3.11.0
Django==5.2.8
django-filter==25.2
djangorestframework==3.14.0
pillow==10.4.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pytz==2025.2
sqlparse==0.5.3
tzdata==2025.2
